moves getUser to utilities
