+++ examples/auth/payload/src/app/(app)/_utilities/getUser.ts
